pyproject.toml
[tool.poetry]
name = "loki"
version = "0.1.0"
description = "Python OpenID Connect Client"
authors = ["SafinWasi <6601566+SafinWasi@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "main" }]

[tool.poetry.dependencies]
python = "^3.10"
oauthlib = {extras = ["signedtoken"], version = "^3.2.2"}
python-dotenv = "^1.0.0"
structlog = "^22.3.0"
requests = "^2.28.2"
requests-oauthlib = "^1.3.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]